SLD Logging: Use fakeBool and merge into main.c
